feature(lint): formatting run.c

+++ src/run.c
@@ -13,7 +13,6 @@
  * limitations under the License.
  */
 
-#include <ctype.h>
 #include <stdio.h>
 #include <string.h>
 #include "../thirdparty/c-logger/src/logger.h"
@@ -128,7 +127,7 @@ int run() {
         while (true) {
             evac_moving_step(graph, zones, transits);
             evac_time_inc();
-            bim_output_body(bim, evac_get_time_m(), fp_detail);
+            bim_output_body(bim, (float)evac_get_time_m(), fp_detail);
 
             double num_of_people = 0;
             for (size_t i = 0; i < zones->length; i++) {
@@ -210,7 +209,7 @@ void applying_scenario_bim_params(bim_t *bim, const bim_cfg_scenario_t *cfg_scen
         if (zone->sign == OUTSIDE) continue;
 
         if (cfg_scenario->distribution.type == distribution_uniform) {
-            bim_tools_set_people_to_zone(zone, (zone->area * cfg_scenario->distribution.density));
+            bim_tools_set_people_to_zone(zone, (float)(zone->area * cfg_scenario->distribution.density));
         }
 
         // A special set up the density of item of bim
@@ -218,7 +217,7 @@ void applying_scenario_bim_params(bim_t *bim, const bim_cfg_scenario_t *cfg_scen
             const special_t special = cfg_scenario->distribution.special[s];
             for (size_t u = 0; u < special.num_of_uuids; u++) {
                 if (uuideq(zone->uuid.x, special.uuid[u].x)) {
-                    bim_tools_set_people_to_zone(zone, (zone->area * special.value));
+                    bim_tools_set_people_to_zone(zone, (float)(zone->area * special.value));
                 }
             }
         }
